requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
dnspython==2.4.2
python-whois==0.7.3
tldextract==3.4.0
rapidfuzz==2.13.7
cryptography==41.0.3
python-dotenv==1.0.0
